+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 Vincent Ohlhauser
 
+## [26/02/10]
+
+### Added
+- Files from Vera, added at project root for later use : 
+  * BSA_manuscript.rmd
+  * phenotype_RIL_II.xlsx
+  * phenotyping_script.R
+
+### Changed
+- _install_packages.R : Re-arranged packages by source (CRAN/Bioconductor/Github)
+- _install_packages.R : Switched to package install by `pak::pkg_install()`, this migh require some fine tuning
+
 ## [26/02/09]
 
 ### Added